Favorite buttons are rendering properly now
